socialnet_faas/functions/video_thumbnail/requirements-test.txt
pytest==8.4.1
moto==5.1.8
Pillow==11.3.0
boto3==1.39.0
